requirements.txt
fastapi[standard]
python-dotenv
aiohttp
uvicorn
pytubefix